refactor packagenames
