+++ core/src/blog/boomerangbeast/factory/BodyFactory.java
@@ -1,4 +1,4 @@
-package blog.boomerangbeast.Factory;
+package blog.boomerangbeast.factory;
 
 import com.badlogic.gdx.math.Vector2;
 import com.badlogic.gdx.physics.box2d.Body;
